refactor: removed dependence of Business in SearchViewModel
Since this algorithm did not need to depend on all the variables of
Business, I simplified the input to be just the parameters that are
needed for the query. Also, I thought it would be easier for
SearchViewModel to return a list of ids of Businesses that match the
query. Therefore, I returned and now SearchViewModel returns ids.